+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changes
 =======
 
+## UNRELEASED
+
+* Feature: Add `formideploy archives` action.
+  [#20](https://github.com/FormidableLabs/formideploy/issues/20)
+
 ## 0.3.1
 
 * Bug: Fix issue for base website and `bucket//archive` empty path issue.